added package system to be more MVC like
