--- lesson05/by.epam.javatraining.shmarlouski.lesson05/src/task01/Main.java
@@ -1,16 +1,0 @@
-package task01;
-
-import static task01.RandomNumber.getRandomInt;
-
-public class Main {
-
-	public static void main(String[] args) {
-
-		int dragonYears = getRandomInt(1000);
-
-		int heads = DragonLogic.getNumberOfHeads(dragonYears);
-		int eyes = DragonLogic.getNumberOfEyes(heads);
-
-		Printer.print("The dragon has " + heads + " heads and " + eyes + " eyes.");
-	}
-}
