+++ lesson05/by.epam.javatraining.shmarlouski.lesson05/src/task01/model/RandomNumber.java
@@ -1,4 +1,4 @@
-package task01;
+package task01.model;
 
 import java.util.Random;
