+++ lesson05/by.epam.javatraining.shmarlouski.lesson05/src/task01/view/Printer.java
@@ -1,0 +1,8 @@
+package task01.view;
+
+public class Printer {
+
+	public static void print(String message) {
+		System.out.println(message);
+	}
+}
