+++ lesson05/by.epam.javatraining.shmarlouski.lesson05/src/task02/model/LetterChecker.java
@@ -1,4 +1,4 @@
-package task02;
+package task02.model;
 
 public class LetterChecker {
