+++ lesson05/by.epam.javatraining.shmarlouski.lesson05/src/task02/view/Printer.java
@@ -1,4 +1,4 @@
-package task01;
+package task02.view;
 
 public class Printer {
