+++ lesson05/by.epam.javatraining.shmarlouski.lesson05/src/task03/controller/Main.java
@@ -1,0 +1,18 @@
+package task03.controller;
+
+import static task03.model.MoodSensor.getMood;
+import static task03.model.RandomNumber.getRandomInt;
+import static task03.view.Printer.print;
+
+
+public class Main {
+
+	public static void main(String[] args) {
+
+		int realFeel = getRandomInt(10);
+
+		String currentMood = getMood(realFeel);
+
+		print(currentMood);
+	}
+}
