+++ lesson05/by.epam.javatraining.shmarlouski.lesson05/src/task03/model/MoodSensor.java
@@ -1,4 +1,4 @@
-package task03;
+package task03.model;
 
 public class MoodSensor {
