+++ lesson05/by.epam.javatraining.shmarlouski.lesson05/src/task03/model/RandomNumber.java
@@ -1,0 +1,12 @@
+package task03.model;
+
+import java.util.Random;
+
+public class RandomNumber {
+	
+	private static Random random = new Random();
+	
+	public static int getRandomInt(int limit) {
+		return random.nextInt(limit);
+	}
+}
