+++ lesson05/by.epam.javatraining.shmarlouski.lesson05/test/task03/MoodSensorTest.java
@@ -1,8 +1,8 @@
 package task03;
 
 import static org.junit.Assert.*;
-
 import org.junit.Test;
+import task03.model.MoodSensor;
 
 public class MoodSensorTest {
